Updates event types
