+++ types/event_record.go
@@ -55,241 +55,34 @@ func (e *EventRecordsRaw) Decode(decoder scale.Decoder) error {
 	return nil
 }
 
-// EventBalancesNewAccount is emitted when a new account was created
-type EventBalancesNewAccount struct {
-	Phase     Phase
-	AccountID AccountID
-	Balance   U128
-	Topics    []Hash
-}
-
-// EventBalancesReapedAccount is emitted when an account was reaped
-type EventBalancesReapedAccount struct {
-	Phase     Phase
-	AccountID AccountID
-	Topics    []Hash
-}
-
-// EventBalancesTransfer is emitted when a transfer succeeded (from, to, value, fees)
-type EventBalancesTransfer struct {
-	Phase  Phase
-	From   AccountID
-	To     AccountID
-	Value  U128
-	Fees   U128
-	Topics []Hash
-}
-
-// EventGrandpaNewAuthorities is emitted when a new authority set has been applied
-type EventGrandpaNewAuthorities struct {
-	Phase          Phase
-	NewAuthorities []struct {
-		AuthorityID     AuthorityID
-		AuthorityWeight U64
-	}
-	Topics []Hash
-}
-
-// EventGrandpaPaused is emitted when the current authority set has been paused
-type EventGrandpaPaused struct {
-	Phase  Phase
-	Topics []Hash
-}
-
-// EventGrandpaResumed is emitted when the current authority set has been resumed
-type EventGrandpaResumed struct {
-	Phase  Phase
-	Topics []Hash
-}
-
-// EventImOnlineAllGood is emitted when at the end of the session, no offence was committed
-type EventImOnlineAllGood struct {
-	Phase  Phase
-	Topics []Hash
-}
-
-// EventImOnlineHeartbeatReceived is emitted when a new heartbeat was received from AuthorityId
-type EventImOnlineHeartbeatReceived struct {
-	Phase       Phase
-	AuthorityID AuthorityID
-	Topics      []Hash
-}
-
-// Exposure lists the own and nominated stake of a validator
-type Exposure struct {
-	Total  UCompact
-	Own    UCompact
-	Others []IndividualExposure
-}
-
-// IndividualExposure contains the nominated stake by one specific third party
-type IndividualExposure struct {
-	Who   AccountID
-	Value UCompact
-}
-
-// EventImOnlineSomeOffline is emitted when the end of the session, at least once validator was found to be offline
-type EventImOnlineSomeOffline struct {
-	Phase                Phase
-	IdentificationTuples []struct {
-		ValidatorID        AccountID
-		FullIdentification Exposure
-	}
-	Topics []Hash
-}
-
-// EventIndicesNewAccountIndex is emitted when a new account index was assigned. This event is not triggered
-// when an existing index is reassigned to another AccountId
-type EventIndicesNewAccountIndex struct {
-	Phase        Phase
-	AccountID    AccountID
-	AccountIndex AccountIndex
-	Topics       []Hash
-}
-
-// EventOffencesOffence is emitted when there is an offence reported of the given kind happened at the session_index
-// and (kind-specific) time slot. This event is not deposited for duplicate slashes
-type EventOffencesOffence struct {
-	Phase          Phase
-	Kind           Bytes16
-	OpaqueTimeSlot Bytes
-	Topics         []Hash
-}
-
-// EventSessionNewSession is emitted when a new session has happened. Note that the argument is the session index,
-// not the block number as the type might suggest
-type EventSessionNewSession struct {
-	Phase        Phase
-	SessionIndex U32
-	Topics       []Hash
-}
-
-// EventStakingOldSlashingReportDiscarded is emitted when an old slashing report from a prior era was discarded because
-// it could not be processed
-type EventStakingOldSlashingReportDiscarded struct {
-	Phase        Phase
-	SessionIndex U32
-	Topics       []Hash
-}
-
-// EventStakingReward is emitted when all validators have been rewarded by the first balance; the second is the
-// remainder, from the maximum amount of reward.
-type EventStakingReward struct {
-	Phase     Phase
-	Balance   U128
-	Remainder U128
-	Topics    []Hash
-}
-
-// EventStakingSlash is emitted when one validator (and its nominators) has been slashed by the given amount
-type EventStakingSlash struct {
-	Phase     Phase
-	AccountID AccountID
-	Balance   U128
-	Topics    []Hash
-}
-
-// EventSystemExtrinsicSuccessV8 is emitted when an extrinsic completed successfully
-//
-// Deprecated: EventSystemExtrinsicSuccessV8 exists to allow users to simply implement their own EventRecords struct if
-// they are on metadata version 8 or below. Use EventSystemExtrinsicSuccess otherwise
-type EventSystemExtrinsicSuccessV8 struct {
-	Phase  Phase
-	Topics []Hash
-}
-
-// EventSystemExtrinsicSuccess is emitted when an extrinsic completed successfully
-type EventSystemExtrinsicSuccess struct {
-	Phase        Phase
-	DispatchInfo DispatchInfo
-	Topics       []Hash
-}
-
-// DispatchInfo contains a bundle of static information collected from the `#[weight = $x]` attributes.
-type DispatchInfo struct {
-	// Weight of this transaction
-	Weight U32
-	// Class of this transaction
-	Class DispatchClass
-	/// PaysFee indicates whether this transaction pays fees
-	PaysFee bool
-}
-
-// DispatchClass is a generalized group of dispatch types. This is only distinguishing normal, user-triggered
-// transactions (`Normal`) and anything beyond which serves a higher purpose to the system (`Operational`).
-type DispatchClass struct {
-	// A normal dispatch
-	IsNormal bool
-	// An operational dispatch
-	IsOperational bool
-}
-
-func (d *DispatchClass) Decode(decoder scale.Decoder) error {
-	b, err := decoder.ReadOneByte()
-	if b == 0 {
-		d.IsNormal = true
-	} else if b == 1 {
-		d.IsOperational = true
-	}
-	return err
-}
-
-func (d DispatchClass) Encode(encoder scale.Encoder) error {
-	var err error
-	if d.IsNormal {
-		err = encoder.PushByte(0)
-	} else if d.IsOperational {
-		err = encoder.PushByte(1)
-	}
-	return err
-}
-
-// EventSystemExtrinsicFailedV8 is emitted when an extrinsic failed
-//
-// Deprecated: EventSystemExtrinsicFailedV8 exists to allow users to simply implement their own EventRecords struct if
-// they are on metadata version 8 or below. Use EventSystemExtrinsicFailed otherwise
-type EventSystemExtrinsicFailedV8 struct {
-	Phase         Phase
-	DispatchError DispatchError
-	Topics        []Hash
-}
-
-// EventSystemExtrinsicFailed is emitted when an extrinsic failed
-type EventSystemExtrinsicFailed struct {
-	Phase         Phase
-	DispatchError DispatchError
-	DispatchInfo  DispatchInfo
-	Topics        []Hash
-}
-
-// EventTreasuryDeposit is emitted when some funds have been deposited
-type EventTreasuryDeposit struct {
-	Phase   Phase
-	Balance U128
-	Topics  []Hash
-}
-
 // EventRecords is a default set of possible event records that can be used as a target for
 // `func (e EventRecordsRaw) Decode(...`
 type EventRecords struct {
-	Balances_NewAccount                []EventBalancesNewAccount                //nolint:stylecheck,golint
-	Balances_ReapedAccount             []EventBalancesReapedAccount             //nolint:stylecheck,golint
+	Balances_Endowed                   []EventBalancesEndowed                   //nolint:stylecheck,golint
+	Balances_DustLost                  []EventBalancesDustLost                  //nolint:stylecheck,golint
 	Balances_Transfer                  []EventBalancesTransfer                  //nolint:stylecheck,golint
+	Balances_BalanceSet                []EventBalancesBalanceSet                //nolint:stylecheck,golint
+	Balances_Deposit                   []EventBalancesDeposit                   //nolint:stylecheck,golint
 	Grandpa_NewAuthorities             []EventGrandpaNewAuthorities             //nolint:stylecheck,golint
 	Grandpa_Paused                     []EventGrandpaPaused                     //nolint:stylecheck,golint
 	Grandpa_Resumed                    []EventGrandpaResumed                    //nolint:stylecheck,golint
-	ImOnline_AllGood                   []EventImOnlineAllGood                   //nolint:stylecheck,golint
 	ImOnline_HeartbeatReceived         []EventImOnlineHeartbeatReceived         //nolint:stylecheck,golint
+	ImOnline_AllGood                   []EventImOnlineAllGood                   //nolint:stylecheck,golint
 	ImOnline_SomeOffline               []EventImOnlineSomeOffline               //nolint:stylecheck,golint
-	Indices_NewAccountIndex            []EventIndicesNewAccountIndex            //nolint:stylecheck,golint
+	Indices_IndexAssigned              []EventIndicesIndexAssigned              //nolint:stylecheck,golint
+	Indices_IndexFreed                 []EventIndicesIndexFreed                 //nolint:stylecheck,golint
 	Offences_Offence                   []EventOffencesOffence                   //nolint:stylecheck,golint
 	Session_NewSession                 []EventSessionNewSession                 //nolint:stylecheck,golint
-	Staking_OldSlashingReportDiscarded []EventStakingOldSlashingReportDiscarded //nolint:stylecheck,golint
 	Staking_Reward                     []EventStakingReward                     //nolint:stylecheck,golint
 	Staking_Slash                      []EventStakingSlash                      //nolint:stylecheck,golint
+	Staking_OldSlashingReportDiscarded []EventStakingOldSlashingReportDiscarded //nolint:stylecheck,golint
 	System_ExtrinsicSuccess            []EventSystemExtrinsicSuccess            //nolint:stylecheck,golint
 	System_ExtrinsicFailed             []EventSystemExtrinsicFailed             //nolint:stylecheck,golint
+	System_CodeUpdates                 []EventSystemCodeUpdated                 //nolint:stylecheck,golint
+	System_NewAccount                  []EventSystemNewAccount                  //nolint:stylecheck,golint
+	System_KilledAccount               []EventSystemKilledAccount               //nolint:stylecheck,golint
 	Treasury_Deposit                   []EventTreasuryDeposit                   //nolint:stylecheck,golint
+	// TODO: Add missing treasury events
 }
 
 // DecodeEventRecords decodes the events records from an EventRecordRaw into a target t using the given Metadata m
@@ -354,7 +147,7 @@ func (e EventRecordsRaw) DecodeEventRecords(m *Metadata, t interface{}) error {
 		moduleName, eventName, err := m.FindEventNamesForEventID(id)
 		// moduleName, eventName, err := "System", "ExtrinsicSuccess", nil
 		if err != nil {
-			return fmt.Errorf("unable to find event with EventID %v in metadata for event #%v", id, i)
+			return fmt.Errorf("unable to find event with EventID %v in metadata for event #%v: %s", id, i, err)
 		}
 
 		log.Debug(fmt.Sprintf("event #%v is in module %v with event name %v", i, moduleName, eventName))
