+++ types/events.go
@@ -1,0 +1,280 @@
+// Go Substrate RPC Client (GSRPC) provides APIs and types around Polkadot and any Substrate-based chain RPC calls
+//
+// Copyright 2019 Centrifuge GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package types
+
+import (
+	"github.com/centrifuge/go-substrate-rpc-client/scale"
+)
+
+// EventBalancesEndowed is emitted when an account is created with some free balance
+type EventBalancesEndowed struct {
+	Phase   Phase
+	Who     AccountID
+	Balance U128
+	Topics  []Hash
+}
+
+// EventDustLost is emitted when an account is removed with a balance that is
+// non-zero but below ExistentialDeposit, resulting in a loss.
+type EventBalancesDustLost struct {
+	Phase   Phase
+	Who     AccountID
+	Balance U128
+	Topics  []Hash
+}
+
+// EventBalancesTransfer is emitted when a transfer succeeded (from, to, value)
+type EventBalancesTransfer struct {
+	Phase  Phase
+	From   AccountID
+	To     AccountID
+	Value  U128
+	Topics []Hash
+}
+
+// EventBalanceSet is emitted when a balance is set by root
+type EventBalancesBalanceSet struct {
+	Phase    Phase
+	Who      AccountID
+	Free     U128
+	Reserved U128
+	Topics   []Hash
+}
+
+// EventDeposit is emitted when an account receives some free balance
+type EventBalancesDeposit struct {
+	Phase   Phase
+	Who     AccountID
+	Balance U128
+	Topics  []Hash
+}
+
+// EventGrandpaNewAuthorities is emitted when a new authority set has been applied
+type EventGrandpaNewAuthorities struct {
+	Phase          Phase
+	NewAuthorities []struct {
+		AuthorityID     AuthorityID
+		AuthorityWeight U64
+	}
+	Topics []Hash
+}
+
+// EventGrandpaPaused is emitted when the current authority set has been paused
+type EventGrandpaPaused struct {
+	Phase  Phase
+	Topics []Hash
+}
+
+// EventGrandpaResumed is emitted when the current authority set has been resumed
+type EventGrandpaResumed struct {
+	Phase  Phase
+	Topics []Hash
+}
+
+// EventImOnlineHeartbeatReceived is emitted when a new heartbeat was received from AuthorityId
+type EventImOnlineHeartbeatReceived struct {
+	Phase       Phase
+	AuthorityID AuthorityID
+	Topics      []Hash
+}
+
+// EventImOnlineAllGood is emitted when at the end of the session, no offence was committed
+type EventImOnlineAllGood struct {
+	Phase  Phase
+	Topics []Hash
+}
+
+// Exposure lists the own and nominated stake of a validator
+type Exposure struct {
+	Total  UCompact
+	Own    UCompact
+	Others []IndividualExposure
+}
+
+// IndividualExposure contains the nominated stake by one specific third party
+type IndividualExposure struct {
+	Who   AccountID
+	Value UCompact
+}
+
+// EventImOnlineSomeOffline is emitted when the end of the session, at least once validator was found to be offline
+type EventImOnlineSomeOffline struct {
+	Phase                Phase
+	IdentificationTuples []struct {
+		ValidatorID        AccountID
+		FullIdentification Exposure
+	}
+	Topics []Hash
+}
+
+// EventIndicesIndexAssigned is emitted when an index is assigned to an AccountID.
+type EventIndicesIndexAssigned struct {
+	Phase        Phase
+	AccountID    AccountID
+	AccountIndex AccountIndex
+	Topics       []Hash
+}
+
+// EventIndicesIndexFreed is emitted when an index is unassigned.
+type EventIndicesIndexFreed struct {
+	Phase        Phase
+	AccountIndex AccountIndex
+	Topics       []Hash
+}
+
+// EventOffencesOffence is emitted when there is an offence reported of the given kind happened at the session_index
+// and (kind-specific) time slot. This event is not deposited for duplicate slashes
+type EventOffencesOffence struct {
+	Phase          Phase
+	Kind           Bytes16
+	OpaqueTimeSlot Bytes
+	Topics         []Hash
+}
+
+// EventSessionNewSession is emitted when a new session has happened. Note that the argument is the session index,
+// not the block number as the type might suggest
+type EventSessionNewSession struct {
+	Phase        Phase
+	SessionIndex U32
+	Topics       []Hash
+}
+
+// EventStakingReward is emitted when all validators have been rewarded by the first balance; the second is the
+// remainder, from the maximum amount of reward.
+type EventStakingReward struct {
+	Phase     Phase
+	Balance   U128
+	Remainder U128
+	Topics    []Hash
+}
+
+// EventStakingSlash is emitted when one validator (and its nominators) has been slashed by the given amount
+type EventStakingSlash struct {
+	Phase     Phase
+	AccountID AccountID
+	Balance   U128
+	Topics    []Hash
+}
+
+// EventStakingOldSlashingReportDiscarded is emitted when an old slashing report from a prior era was discarded because
+// it could not be processed
+type EventStakingOldSlashingReportDiscarded struct {
+	Phase        Phase
+	SessionIndex U32
+	Topics       []Hash
+}
+
+// EventSystemExtrinsicSuccessV8 is emitted when an extrinsic completed successfully
+//
+// Deprecated: EventSystemExtrinsicSuccessV8 exists to allow users to simply implement their own EventRecords struct if
+// they are on metadata version 8 or below. Use EventSystemExtrinsicSuccess otherwise
+type EventSystemExtrinsicSuccessV8 struct {
+	Phase  Phase
+	Topics []Hash
+}
+
+// EventSystemExtrinsicSuccess is emitted when an extrinsic completed successfully
+type EventSystemExtrinsicSuccess struct {
+	Phase        Phase
+	DispatchInfo DispatchInfo
+	Topics       []Hash
+}
+
+// DispatchInfo contains a bundle of static information collected from the `#[weight = $x]` attributes.
+type DispatchInfo struct {
+	// Weight of this transaction
+	Weight U32
+	// Class of this transaction
+	Class DispatchClass
+	/// PaysFee indicates whether this transaction pays fees
+	PaysFee bool
+}
+
+// DispatchClass is a generalized group of dispatch types. This is only distinguishing normal, user-triggered
+// transactions (`Normal`) and anything beyond which serves a higher purpose to the system (`Operational`).
+type DispatchClass struct {
+	// A normal dispatch
+	IsNormal bool
+	// An operational dispatch
+	IsOperational bool
+}
+
+func (d *DispatchClass) Decode(decoder scale.Decoder) error {
+	b, err := decoder.ReadOneByte()
+	if b == 0 {
+		d.IsNormal = true
+	} else if b == 1 {
+		d.IsOperational = true
+	}
+	return err
+}
+
+func (d DispatchClass) Encode(encoder scale.Encoder) error {
+	var err error
+	if d.IsNormal {
+		err = encoder.PushByte(0)
+	} else if d.IsOperational {
+		err = encoder.PushByte(1)
+	}
+	return err
+}
+
+// EventSystemExtrinsicFailedV8 is emitted when an extrinsic failed
+//
+// Deprecated: EventSystemExtrinsicFailedV8 exists to allow users to simply implement their own EventRecords struct if
+// they are on metadata version 8 or below. Use EventSystemExtrinsicFailed otherwise
+type EventSystemExtrinsicFailedV8 struct {
+	Phase         Phase
+	DispatchError DispatchError
+	Topics        []Hash
+}
+
+// EventSystemExtrinsicFailed is emitted when an extrinsic failed
+type EventSystemExtrinsicFailed struct {
+	Phase         Phase
+	DispatchError DispatchError
+	DispatchInfo  DispatchInfo
+	Topics        []Hash
+}
+
+// EventSystemCodeUpdated is emitted when the runtime code (`:code`) is updated
+type EventSystemCodeUpdated struct {
+	Phase  Phase
+	Topics []Hash
+}
+
+// EventSystemNewAccount is emitted when a new account was created
+type EventSystemNewAccount struct {
+	Phase   Phase
+	Who     AccountID
+	Balance U128
+	Topics  []Hash
+}
+
+// EventSystemKilledAccount is emitted when an account is reaped
+type EventSystemKilledAccount struct {
+	Phase  Phase
+	Who    AccountID
+	Topics []Hash
+}
+
+// EventTreasuryDeposit is emitted when some funds have been deposited
+type EventTreasuryDeposit struct {
+	Phase   Phase
+	Balance U128
+	Topics  []Hash
+}
